Add DataStore and Interactions
DataStoreInteractions are data that is related to a message or interaction that needs to be retrieved when a user responds to a message. 

Notes:
DataStoreInteractions will auto be retrieved using the original channel and ts of the message the user is responding to.

In most cases the response from the plugin is the json response from slack of sending a message. This data will be autolinked to the new_interaction object in the request. If there is no new_interaction set by the plugin, it will be skipped. It can also be skipped by setting request.auto_link to False. 

You can create more than one interaction entry per request, it will just have to be done manually using the functions added to the request module.

TODO:
Need to add in a function to look for all interactions that have not been followed up and that the followup_ts has passed. In these cases it should take the bot and followup_action and execute them. 

Also we may need to purge old interactions in the datastore every once in a while